api/requirements.txt
fastapi
uvicorn
supabase
pandas
numpy
tushare
python-dotenv
requests
pydantic
pydantic-settings
email-validator
httpx
dashscope
python-multipart
